requirements/base.txt
django==2.1.1
django-attachments>=1.3
django-contrib-comments
django-glrm
django-grappelli
django-modern-rpc
django-simple-history
django-uuslug
django-vinaigrette
# because of https://github.com/pycontribs/jira/issues/501
jira==1.0.10
Markdown
python-bugzilla
PyGithub
social-auth-app-django
django-slack
rq==0.12.0
redis==2.10.6
django-rq
dacc==0.0.5
python-dateutil
django-adminplus